Removed the callback component
