+++ 01-Login/src/app/app-routing.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './containers/home/home.component';
-import { CallbackComponent } from './containers/callback/callback.component';
 import { ProfileComponent } from './containers/profile/profile.component';
 import { LoginGuard } from './auth/login.guard';
 
@@ -9,10 +8,6 @@ const routes: Routes = [
   {
     path: '',
     component: HomeComponent
-  },
-  {
-    path: 'callback',
-    component: CallbackComponent
   },
   {
     path: 'profile',
