+++ 01-Login/src/app/components/nav-bar/nav-bar.component.ts
@@ -30,7 +30,7 @@ export class NavBarComponent implements OnInit {
 
   async login() {
     await this.auth0Client.loginWithRedirect({
-      redirect_uri: `${window.location.origin}/callback`
+      redirect_uri: window.location.origin
     });
   }
