--- 02-Calling-an-API/src/app/containers/callback/callback.component.html
@@ -1,1 +1,0 @@
-<app-loading></app-loading>
